backend/requirements.txt
fastapi
uvicorn
pandas
prophet
matplotlib
geopy
supabase
python-dotenv
requests